+++ apps/web/modules/design-system/theme/colors.ts
@@ -25,3 +25,23 @@ export const colors = {
     'red-02': '#FFEEEB',
   },
 } as const;
+
+export const textColors: Record<ColorName, string> = {
+  text: 'text-text',
+  green: 'text-green',
+  orange: 'text-orange',
+  pink: 'text-pink',
+  purple: 'text-purple',
+  white: 'text-white',
+  'grey-02': 'text-grey-02',
+  'grey-04': 'text-grey-04',
+  ctaPrimary: 'text-ctaPrimary',
+  ctaHover: 'text-ctaHover',
+  ctaTertiary: 'text-ctaTertiary',
+  bg: 'text-bg',
+  'grey-01': 'text-grey-01',
+  'grey-03': 'text-grey-03',
+  divider: 'text-divider',
+  'red-01': 'text-red-01',
+  'red-02': 'text-red-02',
+};
