+++ apps/web/modules/design-system/theme/typography.ts
@@ -163,3 +163,31 @@ export const typography = {
     },
   },
 };
+
+export const textStyles: Record<TypographyName, string> = {
+  mainPage: 'text-mainPage',
+  largeTitle: 'text-largeTitle',
+  mediumTitle: 'text-mediumTitle',
+  cardEntityTitle: 'text-cardEntityTitle',
+  smallTitle: 'text-smallTitle',
+  body: 'text-body',
+  tableCell: 'text-tableCell',
+  textLink: 'text-textLink',
+  quote: 'text-quote',
+  listItem: 'text-listItem',
+  button: 'text-button',
+  smallButton: 'text-smallButton',
+  tag: 'text-tag',
+  input: 'text-input',
+  metadata: 'text-metadata',
+  breadcrumb: 'text-breadcrumb',
+  navlink: 'text-navLink',
+  footnote: 'text-footnote',
+  bodySemibold: 'text-bodySemibold',
+  textLinkSemibold: 'text-textLinkSemibold',
+  listSemibold: 'text-listSemibold',
+  quoteMedium: 'text-quoteMedium',
+  inputMedium: 'text-inputMedium',
+  metadataMedium: 'text-metadataMedium',
+  footnoteMedium: 'text-footnoteMedium',
+};
