+++ apps/web/pages/spaces.tsx
@@ -2,7 +2,6 @@ import * as React from 'react';
 import { GetServerSideProps } from 'next';
 import Head from 'next/head';
 import { useLogRocket } from '~/modules/analytics/use-logrocket';
-// import { useAccessControl } from '~/modules/auth/use-access-control';
 import { OboardingCarousel } from '~/modules/components/onboarding-carousel/carousel';
 import { Email } from '~/modules/components/onboarding-carousel/email';
 import { SYSTEM_IDS } from '@geogenesis/ids';
@@ -20,7 +19,6 @@ interface Props {
 
 export default function Spaces({ spaces }: Props) {
   const rootSpaceId = spaces.find(space => space.isRootSpace)?.id ?? '';
-  // const { isEditor, isAdmin } = useAccessControl(rootSpaceId);
   useLogRocket(rootSpaceId);
 
   return (
