--- apps/web/utils/index.ts
@@ -1,50 +1,0 @@
-import { ColorName } from '~/modules/design-system/theme/colors';
-import { TypographyName } from '~/modules/design-system/theme/typography';
-
-export const textColors: Record<ColorName, string> = {
-  text: 'text-text',
-  green: 'text-green',
-  orange: 'text-orange',
-  pink: 'text-pink',
-  purple: 'text-purple',
-  white: 'text-white',
-  'grey-02': 'text-grey-02',
-  'grey-04': 'text-grey-04',
-  ctaPrimary: 'text-ctaPrimary',
-  ctaHover: 'text-ctaHover',
-  ctaTertiary: 'text-ctaTertiary',
-  bg: 'text-bg',
-  'grey-01': 'text-grey-01',
-  'grey-03': 'text-grey-03',
-  divider: 'text-divider',
-  'red-01': 'text-red-01',
-  'red-02': 'text-red-02',
-};
-
-export const textStyles: Record<TypographyName, string> = {
-  mainPage: 'text-mainPage',
-  largeTitle: 'text-largeTitle',
-  mediumTitle: 'text-mediumTitle',
-  cardEntityTitle: 'text-cardEntityTitle',
-  smallTitle: 'text-smallTitle',
-  body: 'text-body',
-  tableCell: 'text-tableCell',
-  textLink: 'text-textLink',
-  quote: 'text-quote',
-  listItem: 'text-listItem',
-  button: 'text-button',
-  smallButton: 'text-smallButton',
-  tag: 'text-tag',
-  input: 'text-input',
-  metadata: 'text-metadata',
-  breadcrumb: 'text-breadcrumb',
-  navlink: 'text-navLink',
-  footnote: 'text-footnote',
-  bodySemibold: 'text-bodySemibold',
-  textLinkSemibold: 'text-textLinkSemibold',
-  listSemibold: 'text-listSemibold',
-  quoteMedium: 'text-quoteMedium',
-  inputMedium: 'text-inputMedium',
-  metadataMedium: 'text-metadataMedium',
-  footnoteMedium: 'text-footnoteMedium',
-};
